[BAD] opcodes/riscv: change how disassembler bytes are formatted
This commit contains a proposal to modify how objdump displays the
instruction bytes when disassembling.  The changes are best explained
with an example.  Here's some current disassembler output:

   101b2:	67e1                	lui	a5,0x18
   101b4:	6dc78513          	addi	a0,a5,1756
   101b8:	2c99                	jal	1040e <puts>

And here's the same snippet with this patch applied:

   101b2:       67e1            lui     a5,0x18
   101b4:       8513 6dc7       addi    a0,a5,1756
   101b8:       2c99            jal     1040e <puts>

I am proposing two changes:

First, I have changed the bytes-per-line setting from 8 to 4, this
reduces the white space that is added.  This does mean that if we
disassemble 6 byte or longer instructions, then the extra bytes will,
by default flow onto the next line.  Here is how the disassembly of a
hypothetical 6 byte instruction would appear:

   10000:       aaaa bbbb       insn   operands, ...
   10004:       cccc

This is exactly how long x86-64 instructions are formatted currently.

How annoying this change is will depend, I think, on how common 6
byte (or longer) instructions are right now.  I've never encountered
any extensions that add such long instructions yet, but I don't keep
that close an eye on all the different extensions.

Or course, there's nothing to stop us increasing the bytes-per-line
later if longer instructions start to become more common.

Additionally, there's always the --wide option for objdump, which
prevents the line wrap behaviour.

The second change that this commit proposes is to change the
bytes-per-chunk to a fixed value of 2.  Currently we dynamically
adjust this value based on the instruction length.  That is why, in
the original output '6dc78513' was formatted as a single chunk, while
in the patched version we have '8513 6dc7' - WHICH IS TOTALLY WRONG!

This is because the bytes are read in little endian order, but are now
only being read in chunks of 2 bytes, hence the first two bytes are
swapped, then the next two bytes are swapped, etc.