gdb: fetch target description during attach on linux
This commit aims to address the problem of attaching to a native Linux
process when the current BFD within GDB (as set by the 'file' command)
is for the wrong architecture.  For example, running on an x86-64
GNU/Linux target, attaching to an x86-64 process, but GDB is given a
riscv-v binary.  Currently, in such a situation I see this assertion:

  (gdb) file /tmp/riscv.exe
  (gdb) attach 3276515
  Attaching to program: /tmp/riscv.exe, process 3276515
  ../../src/gdb/i387-tdep.c:596: internal-error: i387_supply_fxsave: Assertion `tdep->st0_regnum >= I386_ST0_REGNUM' failed.

The assertion is cause because, after attaching, we try to handle the
first stop event.  To do this we try to read registers from the
target, which causes us to call target_ops::fetch_registers, which for
x86-64 leads to amd64_linux_nat_target::fetch_registers, and for my
machine, from there I end up (eventually) in i387_supply_fxsave.

This would all be fine, except, as the current file is a riscv-v
executable, the gdbarch for the current inferior is set to be a riscv-v
gdbarch, and the gdbarch_tdep is a riscv_gdbarch_tdep.  So, in
i387_supply_fxsave we do this:

  i386_gdbarch_tdep *tdep = (i386_gdbarch_tdep *) gdbarch_tdep (arch);

We incorrectly cast the tdep object, and now have undefined behaviour.

The solution to this problem is to ensure that GDB fetches the target
description early enough that we can change the architecture of the
current inferior to be x86-64 before we handle the first stop event.
Currently the target description is fetched in post_create_inferior,
but this is really too late, by this point the first stop even has
been handled, and if the current inferior's architecture was wrong,
then we will have already experienced undefined behaviour by this
point.

And so, in this commit, I propose calling target_find_description from
linux_nat_target::attach.  By this point the target has been attached
too, and is stopped, but GDB has not yet processed the stop,
i.e. tried to read the $pc register from the target.  This is almost
enough to resolve the issue, but there is one additional problem.

I did consider adding the target_find_description call to the
attach_command function, in theory, this would have solved the "needs
to get the target description earlier" problem for all targets,
however, I didn't do this.  The comments in attach_command concerned
me a little, my reading of them was that not every target is
guaranteed to be stopped by the time we return to attach_command, and
I'm pretty sure that calling target_find_description for a target that
is not stopped is a bad idea (as it could require probing for target
features, which might not be allowed on a running target).  And so, I
figure, this might be something that needs handling on a target by
target basis.

Having arranged to fetch the target description earlier, there is just
one problem left to solve.

As part of calling target_find_description we call gdbarch_update_p,
which calls gdbarch_find_by_info, which calls gdbarch_info_fill and
then looks up a suitable gdbarch.

In gdbarch_info_fill, if we have both a target description and a BFD
supplied by the 'file' command, then we will call
choose_architecture_for_target to select which of these architectures
we should use.

The goal of choose_architecture_for_target is to handle the case where
one of the architectures (target description or BFD architecture) is
more specific, we then want to use the more specific architecture.

However, in my case we end up trying to choose between the target
description's x86-64 architecture, and the BFD's riscv-v architecture,
these architectures are not at all compatible, and so, currently,
choose_architecture_for_target returns the risc-v architecture.

This means we continue to try and access the target using the risc-v
gdbarch and riscv_gdbarch_tdep structures, which leads to the same
assertion.

My second proposal in this patch is that, when we are selecting
between two completely incompatible architectures in the function
choose_architecture_for_target, we should default to using the
architecture of the target description.  This means we will then only
try to access the target using a gdbarch that is compatible with the
advertised target description.

With these two fixes in place, now, when I attach to an x86-64 process
but use a riscv-v binary with GDB, I see this:

  (gdb) file /tmp/riscv.exe
  (gdb) attach 3276515
  Attaching to program: /tmp/riscv.exe, process 3276515
  warning: Selected architecture riscv:rv32 is not compatible with reported target architecture i386:x86-64
  Reading symbols from /lib64/libc.so.6...
  (No debugging symbols found in /lib64/libc.so.6)
  Reading symbols from /lib64/ld-linux-x86-64.so.2...
  (No debugging symbols found in /lib64/ld-linux-x86-64.so.2)
  Failed to read a valid object file image from memory.
  0x00007f316710c1e7 in nanosleep () from /lib64/libc.so.6
  (gdb) show architecture
  The target architecture is set to "auto" (currently "i386:x86-64").

The 'Failed to read a valid object file image from memory.' messages
comes from symbol_file_add_from_memory, which is called from
add_vsyscall_page.  I think the problem here is that when we try to
create a BFD to represent the vsysall page we base this BFD off of
GDB's current BFD, which is a risc-v binary in my case.

However, before that error, we do get the warning about the two
architectures being incompatible, and, as you can see GDB has selected
the architecture of the target description.

The debug session is pretty useless at this point as GDB is using the
symbols from one binary while trying to debug a completely different
process, but I think this is fine.  GDB didn't crash, and gave a hint
that there was a problem with the architectures, hopefully, from this
the user can figure out their mistake.

No test as I don't know how I can reliably compile a binary for a
different architecture as part of the testsuite.