objdump: introduce OBJDUMP_COLORS environment variable
Add OBJDUMP_COLORS environment variable, and allow this to control the
default behaviour of objdump when --disassembler-color is not used,
and we are disassembling to a terminal.

After this commit:

  commit a88c79b77036e4778e70d62081c3cfd1044bb8e3
  Date:   Tue Aug 9 14:57:48 2022 +0100

      Default to enabling colored disassembly if output is to a terminal.

If objdump is disassembling to a terminal, and the
--disassembler-color argument is not used, then objdump will apply
styling by default as if --disassembler-color=color was used.

However, some users might not like this color on by default approach,
or might prefer to use --disassembler-color=extended-color by default.

With this commit objdump will use the OBJDUMP_COLORS environment
variable to control its behaviour.

With:

  OBJDUMP_COLORS=off

The default behaviour is --disassembler-color=off, that is no colors
are added unless the user explicitly uses --disassembler-color at the
command line.

With:

  OBJDUMP_COLORS=color

Only the basic colors are used, as if --disassembler-color=color was
specified.  This can be overridden by the user at the command line by
explicitly passing --disassembler-color again.

With:

  OBJDUMP_COLORS=extended

Only the extended 256-colors are used, as if
--disassembler-color=extended-color was specified.  This can be
overridden by the user at the command line by explicitly passing
--disassembler-color again.

	* doc/binutils.texi (objdump): Add an @xref to the
	--disassembler-color description.  Add a new section describing
	how to use OBJDUMP_COLORS environment variable.
	* objdump.c (objdump_colors_var): New global.
	(objdump_default_disassembler_color_mode): New function.
	(main): Use objdump_default_disassembler_color_mode.