objdump: allow the disassembler colors to be customized
This commit extends the use of OBJDUMP_COLORS to allow all the
disassembler colors to be controlled by the user.

The approach used is to have the environment variable OBJDUMP_COLORS
contain a simple map of a two letter code to the appropriate escape
sequence to use. For example:

  OBJDUMP_COLORS="sy=38;5;162:rg=38;5;215"

Makes use of color 162 for symbol names (sy) and color 215 for
register names (rg).  Both of these colors are from the 256 color
space.

Any styles not mentioned in OBJDUMP_COLORS will use their default
colors.

Building on the previous commit, OBJDUMP_COLORS can still control the
default styling mode, for example:

  OBJDUMP_COLORS="extended:sy=38;5;162:rg=38;5;215"

With the addition of the 'extended:' prefix, unless told otherwise by
a command line argument, objdump will use the 256 extended-colors by
default for styling, and symbol names and register names will use the
modified colors.

I don't know if this style of environment encoded settings will be
acceptable or not.  The alternative would be some kind of
configuration file system.  My inspiration for the approach I propose
here is the LS_COLORS variable used by `ls`.

	* doc/binutils.texi (objdump): Extend the description of
	OBJDUMP_COLORS environment variable.
	* objdump.c (disasm_styles): New global.
	(objdump_find_disasm_styles_entry): New function.
	(objdump_fill_styles_array): New function.
	(objdump_color_for_disassembler_style): Complete rewrite, return
	strings from disasm_styles array.
	(objdump_styled_sprintf): Update to account for changes to
	objdump_color_for_disassembler_style.