gdb: rewrite gdb.base/info-types.exp.tcl
This patch was inspired by this proposed change:

  https://sourceware.org/pipermail/gdb-patches/2022-July/190926.html

the goal of that change was to make the info-types tests pass when
using Clang as a compiler.  The difference between g++ and Clang, is
that g++ will give a (made up) name to an anonymous struct or union
that is used in a typedef, Clang doesn't.  The solution proposed in
the patch above is to create a third test pattern to be used for
C++/Clang, at this point the test script seemed to be getting a little
bloated.

In this commit I rewrite the test.  Instead of running the 'info
types' command and expecting a specific set of lines in a particular
order, we now read all the lines from the command output, and place
the lines into a dictionary.  We can then query the dictionary to
check the all the expected lines are present.

Checking for each output line independently has a couple of
advantages:

 1. We no longer care about the output order.  I don't think the
 ordering is particularly important for this test, so long as all the
 expected lines are present,

 2. For lines that are present in both C and C++ we can now share the
 code to check these patterns, I think this makes it clearer what is
 going on, and

 3. It is easy to place a few lines within an `if` block, and so, only
 check for those lines when using a specific compiler.

I've check gdb.base/info-types-c.exp and gdb.base/info-types-c++.exp
using both GCC and Clang.