gdb: some process_stratum_target should not be shared
When multi-target support was added to GDB, an assumption was made
that all process_stratum_target sub-classes could be shared by
multiple inferiors.

For things like the Linux and FreeBSD native targets, as well as the
remote target, this is absolutely true (or was made true).  But some
targets were never updated to be shareable, for example, the
core_target, which is used when reading core-files, stores some of its
state in the program_space, but also, the core-file and the executable
being debugged are closely related.

As each add-inferior call creates an inferior with a new
program_space, and doesn't automatically copy the executable, or the
current core-file, I don't think it really makes sense to "share"
core_target objects between inferiors.

Consider this session:

  $ gdb -q
  (gdb) file test1
  Reading symbols from test1...
  (gdb) core-file core.test1.433190
  [New LWP 433190]
  Core was generated by `./test1'.
  Program terminated with signal SIGSEGV, Segmentation fault.
  #0  0x0000000000401111 in foo () at test1.c:6
  6	  return *global_ptr;
  (gdb) add-inferior
  [New inferior 2]
  Added inferior 2 on connection 1 (core)
  (gdb) info inferiors
    Num  Description       Connection           Executable
  * 1    process 433190    1 (core)             /tmp/multi-core/test1
    2    <null>            1 (core)
  (gdb) info connections
    Num  What  Description
  * 1    core  Local core dump file
  (gdb) inferior 2
  [Switching to inferior 2 [<null>] (<noexec>)]
  (gdb) file test2
  Reading symbols from test2...
  (gdb) core-file core.test2.433203
  [New LWP 433203]
  Core was generated by `./test2'.
  Program terminated with signal SIGSEGV, Segmentation fault.
  #0  0x0000000000401111 in main () at test2.c:6
  6	  return *global_ptr;
  (gdb) info inferiors
    Num  Description       Connection           Executable
    1    process 433190    1 (core)             /tmp/multi-core/test1
  * 2    process 433203    2 (core)             /tmp/multi-core/test2
  (gdb) info connections
    Num  What  Description
    1    core  Local core dump file
  * 2    core  Local core dump file
  (gdb)

After the 'add-inferior' the core_target connection is shared between
the inferiors.  However, as soon as the user sets up the core-file and
executable in the new inferior a new core connection has been created.

I think this behaviour might be confusing, so I'd like to have GDB not
initially share the core connection.  Instead, when the user tries to
add the new inferior a warning is given, and the new inferior is
created without a connection, like this:

  $ gdb -q
  (gdb) file test1
  Reading symbols from test1...
  (gdb) core-file core.test1.433190
  [New LWP 433190]
  Core was generated by `./test1'.
  Program terminated with signal SIGSEGV, Segmentation fault.
  #0  0x0000000000401111 in foo () at test1.c:6
  6	  return *global_ptr;
  (gdb) add-inferior
  [New inferior 2]
  warning: can't share connection 1 (core) between inferiors
  Added inferior 2
  (gdb) info inferiors
    Num  Description       Connection           Executable
  * 1    process 433190    1 (core)             /tmp/multi-core/test1
    2    <null>
  (gdb)

If the user explicitly asks for the new inferior to be created without
a connection, then no warning will be given.

At this point the user is free to setup inferior 2 with a different
executable and core file (or to do anything they like with the
inferior).

In an earlier version of this patch I had GDB error instead of giving
a warning.  However, the error message ended up being something like:

  can't share connection ..... between inferiors, use -no-connection
      option to create an inferior without sharing a connection.

but it seemed better to just create the inferior.

I've updated the docs, and added a NEWS entry for the new warning.  In
the docs for clone-inferior I've added reference to -no-connection,
which was previously missing.