gdb/testsuite: skip gdb.cp/anon-struct.exp when using clang
When clang compiles anonymous structures, it does not add linkage names in
their dwarf representations. This is compounded by clang not adding linkage
names to subprograms of those anonymous structs (for instance, the
constructor).

Since this isn't a bug on clang or GDB, but there is no way to make
anon-struct.exp work when using clang, just mark that test as untested.

Since I was already touching the file, I also added a comment at the top
of the file explaining what it is testing for.