sim/ppc: initialize a memory buffer in all cases
In the ppc simulator's do_fstat function, which provides the fstat
call for the simulator, if the fstat is going to fail then we current
write an uninitialized buffer into the inferior.

In theory, I think this is fine, we also write the error status into
the simulated target, so, given that the fstat has failed, the target
shouldn't be relying on the buffer contents.

However, writing an uninitialized buffer means we might leak simulator
private data into the simulated target, which is probably a bad
thing.  Plus it probably makes life easier if something consistent,
like all zeros, is written rather than random junk, that might look
like a successful call.

So, in this commit, I clear the stat buffer to zero before writing it
into the simulated target, if the fstat call is not run (due to a bad
file descriptor).

Another option would be to just not write the buffer into the
inferior (rather than zeroing it, and writing all zeros).  This would
solve the problem of copying simulator data into the target, but I
think the all zeros will make debugging things in the simulator
easier.