gdb: handle loading shared libraries from /proc/self/fd/
Bug PR gdb/29586 describes a situations where a shared library is
created in memory, then written to a memory mapped file.  The memory
mapped file will show up as a file descriptor within /proc/self/fd/,
and this path is then used with dlopen in order to call functions
within the in-memory shared library.

When attempting to debug this GDB hangs.  The problem is that, GDB
stops at the shared-library event, reads the shared library path from
the inferior, which is "/proc/self/fd/<NUM>", and then GDB attempts to
open this file.

Unfortunately, this means GDB tries to open a file within GDB's
/proc/self/fd/ directory, not within the inferior's directory.  In the
case of our hang it turns out that the file descriptor that is opened
is a pipe, and GDB hangs trying to read from the pipe.

However, the behaviour is really just undefined, depending on which
file descriptor the inferior tries to open, GDB will open, or fail to
open, random files within its /proc/self/fd directory.

The solution proposed in this commit is to hook into solib_find_1, and
spot when GDB is looking for any file in /proc/self/, if this is the
case, then the filename is rewritten as /proc/<PID>, where <PID> is
the process-id of the current inferior.

One complexity that we need to consider is how to handle non-canonical
paths, for example, if the user opens /proc/../proc/self/fd/<NUM> then
GDB also needs to be able to spot this.

My solution is to call gdb_realpath_keepfile, this canonicalizes
everything except the filename (the '/proc/../proc/self/fd' part in
the above example).  The problem this then introduces is that the
'self' is replaces with the pid of GDB itself.  I therefore check,
after canonicalization, for a prefix of /proc/<PID OF GDB>, and if
this is the case, I replace this with /proc/<PID OF INFERIOR>.

There is one remaining issue, which I consider very unlikely, and I
don't propose to fix.  What if the inferior actually did try to dlopen
a file from /proc/<PID OF GDB>?  If this happened then GDB will not be
able to distinguish this from the above case and would still replace
the pid of GDB with the pid of the inferior.  However, this seems
really unlikely - why would the inferior be trying to load parts of
GDB?  So I propose to ignore this case until someone can come up with
an actual scenario where that's useful.

The test passes for the unix, native-gdbserver, and
native-extended-gdbserver targets.  When testing with either of the
gdbserver targets, the test is run using the default empty sysroot,
and also using the 'target:' sysroot.

Bug: https://sourceware.org/bugzilla/show_bug.cgi?id=29586

Co-authored-by: Andrew Burgess <aburgess@redhat.com>